Module message_format::icu

ICU Message Format Support

This module provides support for ICU-style message formatting. The ICU Message Format is widely supported in many languages and environments.

This library aims to support all of the ICU Message Format with the exception of the deprecated ChoiceFormat.

The important functionality provided here is the icu::parse function which generates Message from a string.

Syntax

In our implementation of the ICU Message Format, parameters are always named.

Simple Messages

A simple message is just a variable name and is formatted as is:

"Connecting to {host}..."

A message may contain multiple variables:

"Connecting to {application} on {host}..."

Pluralized Messages

Parsing of plural messages is not yet supported.

Select Messages

Parsing of select messages is not yet supported.

Modules

ast
ICU Message Format AST Support

Functions

parse
Parse some text and hopefully return a Message.